requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
Django==5.2.6
django-appconf==1.1.0
django-cryptography==1.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.4.0
firecrawl==4.3.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
psycopg2-binary
gunicorn